Update validation.js
GGSOC"24

I streamlined some repetitive parts, like checking if a name is at least 3 characters long or if an email looks valid. This makes it easier to read and maintain. Also, for the password and login fields, I simplified things so that it's clearer when a field is required or when passwords don't match. Lastly, I tidied up the date of birth check so it directly tells you if the age is at least 10 years, without extra steps.